requirements.txt
appdirs==1.4.4
asgiref==3.3.4
atomicwrites==1.4.0
attrs==20.3.0
distlib==0.3.1
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
filelock==3.0.12
flake8==3.9.0
flake8-import-order==0.18.1
importlib-metadata==3.10.0
importlib-resources==5.1.2
iniconfig==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.2.3
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tox==3.23.0
typing-extensions==3.7.4.3
virtualenv==20.4.3
zipp==3.4.1
